requirements.txt
streamlit
openai
python-dotenv
pandas
scikit-learn
joblib
matplotlib
scikit-learn
sentence-transformers
torch